:#111 Move ApplicationId into it's own namespace
