+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 -----
 
  * #110 Moving from `Devboard` to `DevboardLib` namespace
+ * #111 Move `Application` related VO's to separate namespace (DevboardLib\GitHub\Installation\ApplicationId -> DevboardLib\GitHub\Application\ApplicationId)
